Fixed two bugs with the app and added Camera permission
Added Camera permission to the app when the user is on an iOS device.
Also fixed a bug where the user is able to select the next button
without selecting a pic first which crashes the program.  Also a bug
related to when a user sends a snap and they get back to the root view
controller their email shows up